mostly cosmetics: "for each", small improvements, synchronize on "final", etc.
also:
refactored "combine(IMessageClient client, int maxBytes)" into "combine(MessageClient client)"
-> IMPORTANT: MessageListener changed static ints to enum
-> BUG: OrderedMessage.assignOrderId(..) change ln 71 to: "connectionOrders = orders.get(message.getMessageClient());"
changed test/Chat to not use DynamicMessageAdapter, revert this later on

+++ jgn/trunk/src/com/captiveimagination/jgn/convert/FloatConverter.java
@@ -45,6 +45,6 @@ public class FloatConverter implements Converter {
 	}
 
 	public void get(Object obj, ByteBuffer buffer) throws IllegalArgumentException, IllegalAccessException, InvocationTargetException {
-		buffer.putFloat(((Float)obj).floatValue());
+		buffer.putFloat((Float) obj);
 	}
 }
